requirements.txt
flask==3.0.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.16.4
functions-framework==3.9.1
